update
# Conflicts:
#	.dart_tool/flutter_build/1e4f6db1b82a86c9f62ce582c17ea95b/.filecache
#	.dart_tool/flutter_build/1e4f6db1b82a86c9f62ce582c17ea95b/app.dill
#	.dart_tool/flutter_build/1e4f6db1b82a86c9f62ce582c17ea95b/debug_ios_bundle_flutter_assets.stamp
#	.dart_tool/flutter_build/1e4f6db1b82a86c9f62ce582c17ea95b/flutter_assets.d
#	.dart_tool/flutter_build/1e4f6db1b82a86c9f62ce582c17ea95b/kernel_snapshot.d
#	.dart_tool/flutter_build/1e4f6db1b82a86c9f62ce582c17ea95b/kernel_snapshot.stamp
#	.flutter-plugins-dependencies
#	build/e5af93aeb062c1faf752195ab72478e5.cache.dill.track.dill
#	build/ios/Debug-iphonesimulator/App.framework/App
#	build/ios/Debug-iphonesimulator/App.framework/_CodeSignature/CodeResources
#	build/ios/Debug-iphonesimulator/App.framework/flutter_assets/kernel_blob.bin
#	build/ios/Debug-iphonesimulator/Runner.app/Frameworks/App.framework/App
#	build/ios/Debug-iphonesimulator/Runner.app/Frameworks/App.framework/_CodeSignature/CodeResources
#	build/ios/Debug-iphonesimulator/Runner.app/Frameworks/App.framework/flutter_assets/kernel_blob.bin
#	build/ios/Debug-iphonesimulator/Runner.app/Runner
#	build/ios/Debug-iphonesimulator/Runner.app/_CodeSignature/CodeResources
#	build/ios/Debug-iphonesimulator/libPods-Runner.a
#	build/ios/iphonesimulator/Runner.app/Frameworks/App.framework/App
#	build/ios/iphonesimulator/Runner.app/Frameworks/App.framework/_CodeSignature/CodeResources
#	build/ios/iphonesimulator/Runner.app/Frameworks/App.framework/flutter_assets/kernel_blob.bin
#	build/ios/iphonesimulator/Runner.app/Runner
#	build/ios/iphonesimulator/Runner.app/_CodeSignature/CodeResources
#	lib/main.dart
#	lib/src/presentation/Diary/add_diary_screen.dart
#	lib/src/presentation/Diary/diary_screen.dart
#	lib/src/widget_screen/Auth/Welcome/welcome_screen.dart

+++ lib/src/presentation/HomePage/edit_comment_screen.dart
@@ -1,7 +1,9 @@
 import 'package:diary/src/bloc/editComment_bloc/edit_comment_bloc.dart';
 import 'package:diary/src/core/api.dart';
 import 'package:diary/src/core/apiPath.dart';
+import 'package:diary/styles/text_style.dart';
 import 'package:dio/dio.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:nb_utils/nb_utils.dart';
@@ -83,9 +85,37 @@ class _EditCommentState extends State<EditComment> {
       children: [
         ElevatedButton(
             onPressed: () {
-              _bloc.deleteComment();
-              toastDeleteComplete("");
-              Navigator.pop(context);
+              WidgetsBinding.instance.addPostFrameCallback((_) {
+                showCupertinoDialog(
+                    context: context,
+                    builder: (context) {
+                      return CupertinoAlertDialog(
+                        title: const Icon(CupertinoIcons.info_circle),
+                        content: const Text(
+                          'Bạn Muốn Xoá Bình Luận Này?',
+                          textAlign: TextAlign.center,
+                        ),
+                        actions: [
+                          CupertinoDialogAction(
+                            isDefaultAction: true,
+                            onPressed: () => Navigator.pop(context),
+                            child: Text("Huỷ", style: StyleApp.textStyle402()),
+                          ),
+                          CupertinoDialogAction(
+                            isDefaultAction: true,
+                            onPressed: () {
+                              _bloc.deleteComment();
+                              toastDeleteComplete("");
+                              Navigator.pop(context);
+                            },
+                            child:
+                            Text("Đồng ý", style: StyleApp.textStyle401()),
+                          ),
+                        ],
+                      );
+                    });
+              });
+
             },
             child: const Text("Delete")),
         ElevatedButton(
